HW2/requirements.txt
asttokens==2.0.5
Brotli==1.0.9
certifi==2025.1.31
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.0
executing==0.8.3
filelock==3.13.1
fonttools==4.55.3
gmpy2==2.2.1
idna==3.7
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy>=1.26.4,<1.27.0
onnx==1.17.0
packaging==24.2
parso==0.8.4
pexpect==4.8.0
pillow==11.1.0
pip==25.0
platformdirs==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
setuptools==75.8.0
sip==6.10.0
six==1.16.0
stack-data==0.2.0
sympy==1.13.1
tomli==2.0.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
unicodedata2==15.1.0
urllib3==2.3.0
wcwidth==0.2.5
wheel==0.45.1
